one more rename of fitting files before PR
