+++ lcls_tools/common/data_analysis/fit/method_base.py
