+++ tests/unit_tests/lcls_tools/common/data_analysis/fit/test_methods.py
@@ -1,4 +1,4 @@
-from lcls_tools.common.data_analysis.fit.method.gaussian_model import GaussianModel
+from lcls_tools.common.data_analysis.fit.methods import GaussianModel
 import numpy as np
 import unittest
 import os
